feat: extract UserGroupSchemaInitializer.kt
