+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/userGroup/UserGroupSchemaInitializer.kt
@@ -1,4 +1,4 @@
-package io.whozoss.agentos.user
+package io.whozoss.agentos.userGroup
 
 import mu.KLogging
 import org.springframework.boot.ApplicationArguments
@@ -9,11 +9,11 @@ import org.springframework.data.neo4j.core.Neo4jClient
 import org.springframework.stereotype.Component
 
 /**
- * Idempotent Neo4j schema initialiser for [User].
+ * Idempotent Neo4j schema initialiser for [UserGroup].
  *
  * Backfills `removed = false` on any pre-existing nodes that have the property absent
  * (stored as `null` before the field was made non-nullable), then ensures the unique
- * constraint on `externalId` via the `ActiveUser` virtual label.
+ * constraint on `(name, namespaceId)` via the `ActiveUserGroup` virtual label.
  */
 @Component
 @Order(1)
@@ -21,42 +21,34 @@ import org.springframework.stereotype.Component
     "'\${agentos.persistence.mode:in-memory}' == 'neo4j' " +
         "or '\${agentos.persistence.mode:in-memory}' == 'embedded-neo4j'",
 )
-class UserSchemaInitializer(
+class UserGroupSchemaInitializer(
     private val neo4jClient: Neo4jClient,
 ) : ApplicationRunner {
 
     override fun run(args: ApplicationArguments) {
         backfillRemoved()
-        ensureExternalIdUniqueConstraint()
-        ensureExternalIdIndex()
+        ensureNameUniqueConstraint()
     }
 
     private fun backfillRemoved() {
         val backfilled =
             neo4jClient
-                .query("MATCH (u:User) WHERE u.removed IS NULL SET u.removed = false RETURN count(u) AS count")
+                .query("MATCH (g:UserGroup) WHERE g.removed IS NULL SET g.removed = false RETURN count(g) AS count")
                 .fetchAs(Long::class.java)
                 .mappedBy { _, record -> record["count"].asLong() }
                 .one()
                 .orElse(0L)
         if (backfilled > 0L) {
-            logger.info { "[UserSchema] Backfilled removed=false on $backfilled User node(s)" }
+            logger.info { "[UserGroupSchema] Backfilled removed=false on $backfilled UserGroup node(s)" }
         }
     }
 
-    private fun ensureExternalIdUniqueConstraint() {
+    private fun ensureNameUniqueConstraint() {
         neo4jClient.query(
-            "CREATE CONSTRAINT user_external_id_unique IF NOT EXISTS " +
-                "FOR (u:ActiveUser) REQUIRE u.externalId IS UNIQUE",
+            "CREATE CONSTRAINT user_group_name_namespace_unique IF NOT EXISTS " +
+                "FOR (g:ActiveUserGroup) REQUIRE (g.name, g.namespaceId) IS UNIQUE",
         ).run()
-        logger.info { "[UserSchema] Constraint user_external_id_unique ensured" }
-    }
-
-    private fun ensureExternalIdIndex() {
-        neo4jClient.query(
-            "CREATE INDEX user_externalId IF NOT EXISTS FOR (u:User) ON (u.externalId)",
-        ).run()
-        logger.info { "[UserSchema] Index user_externalId ensured" }
+        logger.info { "[UserGroupSchema] Constraint user_group_name_namespace_unique ensured" }
     }
 
     companion object : KLogging()
